git-version: simplify the versioning scheme
Now the version for trunk head is simply rx.y-z.  x.y is the last
release, and z is the number of commits after x.y.

For other commits, the version is rx.y-z-g{sha}.  An abbreviated sha
hash is added to distinguish commits with same x, y, and z on different
branches.

If there are uncommited changes, a "+" is appended to the version,
indicating the version is "unclean".

To make it works correctly, a tag "rx.y" have to be created on the last
commit on trunk of x.y release.