Make stopping the network interfaces more robust. SysV only.
When stopping or rebooting the system, the network boot script is run and looks
at all the /etc/sysconfig/ifconfig.* files. Some of the devices these files
refer to may not be active or even refer to interfaces that are not defined on
the current system. This change ignores non-existant interfaces and any
interfaces that are not UP. This change also avoids bogus warning or error
error messages when the system is shutting down.