Add conditional entities for package versions
Those packages only in sysv book should appear after a "<![ %sysv; ["
conditional. But since the entity must always be defined when
profiling (the revision attribute removes text in the output, not
when parsing), there are the same entities with text "not-used"
after "<![ %systemd; [" conditional. Of course, the preceding is
reversed for packages only in the systemd book.