provide revision-specific wget-list-$(REV) file
It's irrational to force a sysv builder to download systemd, or vice
versa.  But we cannot simply make wget-list specific for revision: IIRC
the wget-list file with all packages in either revision is used by some
scripts on rivendell.  So we keep wget-list as is, and provide a new
wget-list-$(REV) file which only contains the packages for one revision.