dependencies: List bzip2, xz, and zlib as libelf dependencies
libelf.so links to libbz2.so.1.0, liblzma.so.5, and libz.so.1.  libz is
really used, but to me liblzma and libbz2 are "unneeded dependencies":
the building system apparently thinks they are needed by zstd but
actually not.  I don't want to "fix the upstream" so just list them here
anyway.